Fixed compilation warnings reported by Clang
src/gl/scene/gl_sprite.cpp:685:34: warning: '&&' within '||' [-Wlogical-op-parentheses]
src/polyrenderer/scene/poly_sprite.cpp:297:34: warning: '&&' within '||' [-Wlogical-op-parentheses]
src/swrenderer/scene/r_opaque_pass.cpp:975:35: warning: '&&' within '||' [-Wlogical-op-parentheses]
src/sound/mididevices/music_timiditypp_mididevice.cpp:548:30: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]

+++ src/sound/mididevices/music_timiditypp_mididevice.cpp
@@ -545,7 +545,7 @@ bool TimidityPPMIDIDevice::LaunchTimidity ()
 	globfree(&glb);
 
 	int strCount = 1;
-	for (spaceIdx = 0; spaceIdx < CommandLine.Len(); spaceIdx++)
+	for (spaceIdx = 0; spaceIdx < static_cast<int>(CommandLine.Len()); spaceIdx++)
 	{
 		if (CommandLine[spaceIdx] == ' ')
 		{
